We can't test on 2.6 because of how travis works
Travis machines are always ubuntu boxes with py2.7 as system python.
We cannot get rpm bindings on 2.6 because they are native and only
installable with packages. Packages are obviously for 2.7 only (system
version) so even with site packages we can't get it to work.

Also we have tests for queryrpm that assume the bindings are available
(for version comparison).